requirements.txt
keyauth.py==1.2.5
customtkinter==5.2.0
pillow==10.0.0
pyinstaller==6.1.0 